Move test to correct location after merge
