+++ source/Calamari.Tests/KubernetesFixtures/KubernetesContextScriptWrapperLiveFixture.cs
@@ -157,48 +157,6 @@ namespace Calamari.Tests.KubernetesFixtures
 
             return workingFolder;
         }
-        
-        [Test]
-        public void DiscoverKubernetesClusterWithNoValidCredentials()
-        {
-            const string accessKeyEnvVar = "AWS_ACCESS_KEY_ID";
-            const string secretKeyEnvVar = "AWS_SECRET_ACCESS_KEY";
-            var originalAccessKey = Environment.GetEnvironmentVariable(accessKeyEnvVar);
-            var originalSecretKey = Environment.GetEnvironmentVariable(secretKeyEnvVar);
-
-            try
-            {
-                Environment.SetEnvironmentVariable(accessKeyEnvVar, "NotValid");
-                Environment.SetEnvironmentVariable(secretKeyEnvVar, "NotValid");
-                
-                var authenticationDetails = new AwsAuthenticationDetails
-                {
-                    Type = "Aws",
-                    Credentials = new AwsCredentials { Type = "worker" },
-                    Role = new AwsAssumedRole { Type = "noAssumedRole" },
-                    Regions = new []{region}
-                };
-                
-                var serviceMessageCollectorLog = new ServiceMessageCollectorLog();
-                Log = serviceMessageCollectorLog;
-                
-                DoDiscovery(authenticationDetails);
-
-                serviceMessageCollectorLog.ServiceMessages.Should().BeEmpty();
-
-                serviceMessageCollectorLog.MessagesErrorFormatted.Should().BeEmpty();
-
-                serviceMessageCollectorLog.StandardError.Should().BeEmpty();
-
-                serviceMessageCollectorLog.MessagesWarnFormatted.Should()
-                                          .Contain("Unable to authorise credentials, see verbose log for details.");
-            }
-            finally
-            {
-                Environment.SetEnvironmentVariable(accessKeyEnvVar, originalAccessKey);
-                Environment.SetEnvironmentVariable(secretKeyEnvVar, originalSecretKey);
-            }
-        }
     }
 }
 #endif
